Fix warnings
git-svn-id: svn+ssh://svn.yandex.ru/yt/trunk@1076 41d65440-b5be-11dd-afe3-b2e846d9b4f8

+++ yt/ytlib/chunk_client/file_chunk_reader.h
@@ -43,7 +43,7 @@ public:
         // Check alignment.
         YASSERT(
             static_cast<i64>(reinterpret_cast<char*>(infoEnd) - reinterpret_cast<char*>(infoBegin)) ==
-            footer.BlockCount * sizeof (TBlockInfo));
+            static_cast<i64>(footer.BlockCount * sizeof (TBlockInfo)));
 
         File->Seek(footer.BlockInfoOffset, sSet);
         File->Read(infoBegin, footer.BlockCount * sizeof (TBlockInfo));
